+++ tests/JestExt.test.ts
@@ -10,7 +10,12 @@ jest.mock('../src/DebugCodeLens', () => ({
   DebugCodeLensProvider: class MockCodeLensProvider {},
 }));
 jest.mock('os');
-jest.mock('../src/decorations');
+jest.mock('../src/decorations/test-status', () => ({
+  TestStatus: jest.fn(),
+}));
+jest.mock('../src/decorations/inline-error', () => ({
+  default: jest.fn(),
+}));
 
 const update = jest.fn();
 const statusBar = {
@@ -20,18 +25,21 @@ jest.mock('../src/StatusBar', () => ({ statusBar }));
 
 import { JestExt } from '../src/JestExt';
 import { ProjectWorkspace } from 'jest-editor-support';
-import { window, workspace, debug } from 'vscode';
+import { window, workspace, debug, ExtensionContext, TextEditorDecorationType } from 'vscode';
 import { hasDocument, isOpenInMultipleEditors } from '../src/editor';
-import * as decorations from '../src/decorations';
+import { TestStatus } from '../src/decorations/test-status';
 import { updateCurrentDiagnostics } from '../src/diagnostics';
 import { JestProcessManager, JestProcess } from '../src/JestProcessManagement';
 import * as messaging from '../src/messaging';
 import { CoverageMapProvider } from '../src/Coverage';
+import inlineError from '../src/decorations/inline-error';
 
 /* eslint jest/expect-expect: ["error", { "assertFunctionNames": ["expect", "expectItTakesNoAction"] }] */
 
 describe('JestExt', () => {
   const getConfiguration = workspace.getConfiguration as jest.Mock<any>;
+  const StateDecorationsMock = TestStatus as jest.Mock;
+  const context = { asAbsolutePath: (text) => text } as ExtensionContext;
   const workspaceFolder = { name: 'test-folder' } as any;
   let projectWorkspace: ProjectWorkspace;
   const channelStub = { appendLine: jest.fn(), clear: jest.fn(), show: jest.fn() } as any;
@@ -62,7 +70,7 @@ describe('JestExt', () => {
 
     beforeEach(() => {
       sut = new JestExt(
-        null,
+        context,
         workspaceFolder,
         projectWorkspace,
         channelStub,
@@ -76,7 +84,6 @@ describe('JestExt', () => {
 
       sut.canUpdateActiveEditor = jest.fn().mockReturnValueOnce(true);
       sut.debugCodeLensProvider.didChange = jest.fn();
-      ((decorations.failingAssertionStyle as unknown) as jest.Mock<{}>).mockReturnValue({});
       ((sut.testResultProvider.getSortedResults as unknown) as jest.Mock<{}>).mockReturnValueOnce({
         success: [],
         fail: [],
@@ -125,8 +132,11 @@ describe('JestExt', () => {
         debugCodeLens: {},
         enableInlineErrorMessages: true,
       };
+      const expected = { key: 'value' };
+      const failingAssertionStyle = inlineError as jest.Mock;
+      failingAssertionStyle.mockReturnValueOnce(expected);
       const sut = new JestExt(
-        null,
+        context,
         workspaceFolder,
         projectWorkspace,
         channelStub,
@@ -141,10 +151,6 @@ describe('JestExt', () => {
         document: { fileName: 'file.js' },
         setDecorations: jest.fn(),
       };
-      const expected = {};
-      ((decorations.failingAssertionStyle as unknown) as jest.Mock<{}>).mockReturnValueOnce(
-        expected
-      );
       sut.canUpdateActiveEditor = jest.fn().mockReturnValueOnce(true);
       sut.testResultProvider.getSortedResults = jest.fn().mockReturnValueOnce({
         success: [],
@@ -181,7 +187,7 @@ describe('JestExt', () => {
 
       const debugConfiguration = { type: 'dummyconfig' };
       const sut = new JestExt(
-        null,
+        context,
         workspaceFolder,
         projectWorkspace,
         channelStub,
@@ -215,7 +221,7 @@ describe('JestExt', () => {
   describe('onDidCloseTextDocument()', () => {
     const projectWorkspace = new ProjectWorkspace(null, null, null, null);
     const sut = new JestExt(
-      null,
+      context,
       workspaceFolder,
       projectWorkspace,
       channelStub,
@@ -244,7 +250,7 @@ describe('JestExt', () => {
   describe('removeCachedTestResults()', () => {
     const projectWorkspace = new ProjectWorkspace(null, null, null, null);
     const sut = new JestExt(
-      null,
+      context,
       workspaceFolder,
       projectWorkspace,
       channelStub,
@@ -280,7 +286,7 @@ describe('JestExt', () => {
   describe('removeCachedAnnotations()', () => {
     const projectWorkspace = new ProjectWorkspace(null, null, null, null);
     const sut = new JestExt(
-      null,
+      context,
       workspaceFolder,
       projectWorkspace,
       channelStub,
@@ -316,7 +322,7 @@ describe('JestExt', () => {
     const editor: any = {};
     const projectWorkspace = new ProjectWorkspace(null, null, null, null);
     const sut = new JestExt(
-      null,
+      context,
       workspaceFolder,
       projectWorkspace,
       channelStub,
@@ -354,7 +360,7 @@ describe('JestExt', () => {
     beforeEach(() => {
       const projectWorkspace = new ProjectWorkspace(null, null, null, null);
       sut = new JestExt(
-        null,
+        context,
         workspaceFolder,
         projectWorkspace,
         channelStub,
@@ -432,7 +438,7 @@ describe('JestExt', () => {
   describe('toggleCoverageOverlay()', () => {
     it('should toggle the coverage overlay visibility', () => {
       const sut = new JestExt(
-        null,
+        context,
         workspaceFolder,
         projectWorkspace,
         channelStub,
@@ -452,7 +458,7 @@ describe('JestExt', () => {
     it('overrides showCoverageOnLoad settings', () => {
       const settings = { showCoverageOnLoad: true } as any;
       const sut = new JestExt(
-        null,
+        context,
         workspaceFolder,
         projectWorkspace,
         channelStub,
@@ -481,7 +487,7 @@ describe('JestExt', () => {
       const editor: any = {};
 
       const sut = new JestExt(
-        null,
+        context,
         workspaceFolder,
         projectWorkspace,
         channelStub,
@@ -498,7 +504,7 @@ describe('JestExt', () => {
     });
     it('should update both decorators and diagnostics for valid editor', () => {
       const sut = new JestExt(
-        null,
+        context,
         workspaceFolder,
         projectWorkspace,
         channelStub,
@@ -539,7 +545,7 @@ describe('JestExt', () => {
       jest.resetAllMocks();
       const projectWorkspace = new ProjectWorkspace(null, null, null, null);
       sut = new JestExt(
-        null,
+        context,
         workspaceFolder,
         projectWorkspace,
         channelStub,
@@ -574,6 +580,7 @@ describe('JestExt', () => {
     let sut: JestExt;
     const mockEditor: any = { document: { uri: { fsPath: `file://a/b/c.js` } } };
     const emptyTestResults = { success: [], fail: [], skip: [], unknown: [] };
+    const failingAssertionStyle = inlineError as jest.Mock;
 
     const settings: any = {
       debugCodeLens: {},
@@ -589,14 +596,15 @@ describe('JestExt', () => {
 
     beforeEach(() => {
       jest.resetAllMocks();
-      ((decorations.failingItName as unknown) as jest.Mock<{}>).mockReturnValue({ key: 'fail' });
-      ((decorations.passingItName as unknown) as jest.Mock<{}>).mockReturnValue({ key: 'pass' });
-      ((decorations.skipItName as unknown) as jest.Mock<{}>).mockReturnValue({ key: 'skip' });
-      ((decorations.notRanItName as unknown) as jest.Mock<{}>).mockReturnValue({ key: 'notRan' });
-
+      StateDecorationsMock.mockImplementation(() => ({
+        passing: { key: 'pass' } as TextEditorDecorationType,
+        failing: { key: 'fail' } as TextEditorDecorationType,
+        skip: { key: 'skip' } as TextEditorDecorationType,
+        unknown: { key: 'unknown' } as TextEditorDecorationType,
+      }));
       const projectWorkspace = new ProjectWorkspace(null, null, null, null);
       sut = new JestExt(
-        null,
+        context,
         workspaceFolder,
         projectWorkspace,
         channelStub,
@@ -630,7 +638,7 @@ describe('JestExt', () => {
             expect(args[1].length).toBe(1);
             break;
           case 'skip':
-          case 'notRan':
+          case 'unknown':
             expect(args[1].length).toBe(0);
             break;
           default:
@@ -642,17 +650,15 @@ describe('JestExt', () => {
     it('will update inlineError decorator only if setting is enabled', () => {
       const testResults2: any = { success: [], fail: [tr1, tr2], skip: [], unknown: [] };
       const expected = {};
-      ((decorations.failingAssertionStyle as unknown) as jest.Mock<{}>).mockReturnValueOnce(
-        expected
-      );
+      failingAssertionStyle.mockReturnValueOnce(expected);
       sut.updateDecorators(testResults2, mockEditor);
-      expect(decorations.failingAssertionStyle).not.toBeCalled();
+      expect(failingAssertionStyle).not.toBeCalled();
       expect(mockEditor.setDecorations).toHaveBeenCalledTimes(4);
 
       jest.clearAllMocks();
       settings.enableInlineErrorMessages = true;
       sut.updateDecorators(testResults2, mockEditor);
-      expect(decorations.failingAssertionStyle).toHaveBeenCalledTimes(2);
+      expect(failingAssertionStyle).toHaveBeenCalledTimes(2);
       expect(mockEditor.setDecorations).toHaveBeenCalledTimes(6);
     });
   });
@@ -670,7 +676,7 @@ describe('JestExt', () => {
       jest.resetAllMocks();
       const projectWorkspace = new ProjectWorkspace(null, null, null, null);
       sut = new JestExt(
-        null,
+        context,
         workspaceFolder,
         projectWorkspace,
         channelStub,
@@ -710,7 +716,7 @@ describe('JestExt', () => {
       const mockProcess: any = mockJestProcess();
 
       const jestExt = new JestExt(
-        null,
+        context,
         workspaceFolder,
         projectWorkspace,
         channelStub,
@@ -794,7 +800,7 @@ describe('JestExt', () => {
       jest.resetAllMocks();
       const projectWorkspace = new ProjectWorkspace(null, null, null, null);
       sut = new JestExt(
-        null,
+        context,
         workspaceFolder,
         projectWorkspace,
         channelStub,
@@ -836,7 +842,7 @@ describe('JestExt', () => {
       }));
       const coverageCodeLensProvider: any = { coverageChanged: jest.fn() };
       const sut = new JestExt(
-        null,
+        context,
         workspaceFolder,
         projectWorkspace,
         channelStub,